# openapi.yaml
openapi: 3.1.0
info:
  title: Guard API — Антиплагиат БГУИР
  version: "1.0.0"
  description: |
    REST API веб-приложения Guard (антиплагиат БГУИР).

    **Аудитории**
    - Браузерный UI (`index.html` / `app.html` / `admin.html`) — cookie `guard_session`.
    - Плагин Moodle СЭО — заголовок `X-API-Key` (= `MOODLE_API_KEY`, ≥16 символов)
      или `Authorization: Bearer`. Логин студента: поле `username`, query `username`
      или заголовок `X-Moodle-Username`.
    - QR-справки — HMAC-подпись `sig` (`REPORT_ACCESS_SECRET`).
    - Ночной cron — `X-Cron-Secret` (= `CLEANUP_CRON_SECRET`).

    **Эндпоинты плагина Moodle:** `POST /api/upload`, `GET /api/documents/{documentId}/analysis`,
    `GET /api/report/{documentId}/links`.

    **Особенности**
    - `POST /api/auth/register` и `POST /api/report` в текущей сборке без авторизации.
    - `GET /api/analysis/health` требует секрет только если задан `CLEANUP_CRON_SECRET`.
  contact:
    name: БГУИР
servers:
  - url: /
    description: Тот же хост, что и сайт Guard
...
paths:
...
  /api/upload:
    post:
      tags: [documents, moodle]
      summary: Загрузка файла и постановка в очередь анализа
      description: |
        Возвращает `202` сразу. ML выполняет `analysis-worker`.
        Плагин Moodle передаёт `X-API-Key` и `username` (LDAP uid).
      security:
        - cookieAuth: []
        - moodleApiKey: []
        - moodleBearer: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, title, content]
              properties:
                file: { type: string, format: binary }
                title: { type: string }
                content: { type: string, description: Извлечённый текст ≥50 символов }
                category: { type: string, example: lab }
                username: { type: string }
                fullName: { type: string }
                institution: { type: string }
                institutionId: { type: string }
                document_type: { type: string, enum: [word, pdf] }
      responses:
        "202":
          description: Поставлено в очередь
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  status: { type: string, example: processing }
                  jobId: { type: integer }
                  document:
                    type: object
                    properties:
                      id: { type: integer }
                      title: { type: string }
                      filename: { type: string }
                      wordCount: { type: integer }
                      category: { type: string }
                      status: { type: string }
                      localPlagiarismPercent: { type: number }
                  message: { type: string }
        "400": { $ref: "#/components/responses/Error" }
        "401": { $ref: "#/components/responses/Error" }
        "409":
          description: У пользователя уже есть активная задача
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean, example: false }
                  error: { type: string }
                  code: { type: string, example: ACTIVE_JOB_EXISTS }
        "503": { $ref: "#/components/responses/Error" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: guard_session
      description: HMAC-сессия после POST /api/auth/login
    moodleApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Равен MOODLE_API_KEY (≥16 символов)
    moodleBearer:
      type: http
      scheme: bearer
      description: Тот же секрет, что MOODLE_API_KEY
    cronSecret:
      type: apiKey
      in: header
      name: X-Cron-Secret
      description: Равен CLEANUP_CRON_SECRET
    qrSignature:
      type: apiKey
      in: query
      name: sig
      description: HMAC подпись REPORT_ACCESS_SECRET
...
  schemas:
...
    ErrorBody:
      type: object
      properties:
        success: { type: boolean, example: false }
        error: { type: string }
        code: { type: string }
...
  responses:
...
    Error:
      description: Ошибка
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorBody" }